categorical sample

+++ keanu-project/src/main/java/io/improbable/keanu/tensor/intgr/Nd4jIntegerTensor.java
@@ -29,7 +29,7 @@ public class Nd4jIntegerTensor implements IntegerTensor {
         return new Nd4jIntegerTensor(values, shape);
     }
 
-    public static Nd4jIntegerTensor create(double value, int[] shape) {
+    public static Nd4jIntegerTensor create(int value, int[] shape) {
         return new Nd4jIntegerTensor(Nd4j.valueArrayOf(shape, value));
     }
